Refactor: Extension 폴더 생성
