+++ PJ2T12_Again12/PJ2T12_Again12/Extensions/Color.swift
@@ -1,0 +1,21 @@
+//
+//  Color.swift
+//  PJ2T12_Again12
+//
+//  Created by KHJ on 2023/12/11.
+//
+
+import Foundation
+import SwiftUI
+
+extension Color {
+    init(hex: UInt, alpha: Double = 1) {
+        self.init(
+            .sRGB,
+            red: Double((hex >> 16) & 0xff) / 255,
+            green: Double((hex >> 08) & 0xff) / 255,
+            blue: Double((hex >> 00) & 0xff) / 255,
+            opacity: alpha
+        )
+    }
+}
