+++ PJ2T12_Again12/PJ2T12_Again12/Views/HomeView.swift
@@ -6,35 +6,7 @@
 //
 
 import SwiftUI
-// TODO: UserModel 고치기 
-// TODO: Extension Folder 생성
-extension Color {
-    init(hex: UInt, alpha: Double = 1) {
-        self.init(
-            .sRGB,
-            red: Double((hex >> 16) & 0xff) / 255,
-            green: Double((hex >> 08) & 0xff) / 255,
-            blue: Double((hex >> 00) & 0xff) / 255,
-            opacity: alpha
-        )
-    }
-}
-// TODO: ViewModifiers Folder 생성
-struct TodoCellModifier: ViewModifier {
-    let status: Bool
-    let hexCode: UInt
-    func body(content: Content) -> some View {
-        content
-            .padding()
-            .bold()
-            .foregroundColor(status ? .white : Color(hex: hexCode))
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(
-                RoundedRectangle(cornerRadius: 20, style: .continuous)
-                    .fill(status ? Color(hex: hexCode, alpha: 0.4) : .white)
-            )
-    }
-}
+
 // TODO: 이모티콘에 대한 고민 나눠보기
 struct HomeView: View {
     @StateObject private var homeVM = HomeViewModel()
@@ -146,6 +118,22 @@ struct HomeView: View {
                 NavigationLink("일정 달성") { DetailView() }
             }
         }
+    }
+}
+
+struct TodoCellModifier: ViewModifier {
+    let status: Bool
+    let hexCode: UInt
+    func body(content: Content) -> some View {
+        content
+            .padding()
+            .bold()
+            .foregroundColor(status ? .white : Color(hex: hexCode))
+            .frame(maxWidth: .infinity, alignment: .leading)
+            .background(
+                RoundedRectangle(cornerRadius: 20, style: .continuous)
+                    .fill(status ? Color(hex: hexCode, alpha: 0.4) : .white)
+            )
     }
 }
